Align destructuring with ES6 draft
Destructuring now supports initializers. For example:

  var {x = 42} = {};
  assertEquals(42, x);

Another change is that we don't use a look ahead parser any more. In the cases
where an assignment pattern is allowed we "transform" array and object literals
into a patterns. This "transformation" is done by reparsing the code.

We also now correctly handle number and string property names.

Destructuring in rest position is not supported in the ES6 draft so that was removed.

BUG=http://code.google.com/p/traceur-compiler/issues/detail?id=95
TEST=test/feature/Destructuring/*

Review URL: https://codereview.appspot.com/6209058