Simplify output of default quasis.
We now generate something as simple as:

`a${ b }c${ 1 + 2 }${ 3 * 4 }`

=>

("a" + b + "c" + (1 + 2) + 3 * 4)

This removes the call to traceur.runtime.defaultQuasi. It also removes empty
strings and correctly wraps expressions in parens as needed.

This patch also fixes a bug where comma expression in the substitution was not
working because it was not wrapped in a paren expression.

BUG=https://code.google.com/p/traceur-compiler/issues/detail?id=107
TEST=test/feature/QausiLiteral/*

Review URL: https://codereview.appspot.com/6352043