Add option to inline external modules for the file compiler.
This needs a lot of cleanup but it works already.

Usage:

node src/filecompiler.js --inline-modules path/to/js-file-with-import.js

This outputs out/path/to/js-file-with-import.compiled.js and
out/path/to/js-file-with-import.map as well as copies all the original files so
that source mapping works correcty.

BUG=https://code.google.com/p/traceur-compiler/issues/detail?id=87
TEST=None

Review URL: https://codereview.appspot.com/6441130