Fix validation of MemberExpression and NewExpression operands
Fixes validation error when the operand is not a LeftExpression, for example:
"new (a || b)()" and "(a || b)[i]".

The correctness of this fix depends on the correctness of the first validation
section of visitCallExpression, since it has essentially been cloned into
visitNewExpression and visitMemberLookupExpression.

My rationale is that anything that could potentially resolve to a function (and
therefore is valid in a CallExpression) could also be called with "new" or
dereferenced with "[ ]".

BUG=http://code.google.com/p/traceur-compiler/issues/detail?id=129
TEST=test/unit/syntax/parsetreevalidator_test.html

Review URL: https://codereview.appspot.com/6641044
Patch from Ben Chan <usrbincc@yahoo.com>.