Optimize transformAny and visitAny
Before this we always computed the method name to call and did a member lookup
based on this. Since transformAny and visitAny are such bottle necks this caused
serious slow down.

Instead we now generate visit and transform methods for all ParseTree classes. The truth for the parse trees are now in a json file and based on this json file we generate ParseTrees.js and ParseTreeType.js

This makes make build and make test about 30% faster.

BUG=None
TEST=None

Review URL: https://codereview.appspot.com/6843074