Allow typeof to "declare" variable for free-variable-checker.
This allows using typeof to guard agains the free variable checker to mark a
variable as not being declared.

if (typeof name !== 'undefined')
  name;

This only tests for typeof IdentifierExpression so it does not guarantee
anything but it does allow us to use the free variable checker in more places.

BUG=https://code.google.com/p/traceur-compiler/issues/detail?id=169
TEST=test/feature/FreeVariableChecker/Typeof.js

Review URL: https://codereview.appspot.com/6854061