This changes how we scan and parts template literals.
We now have 4 different token types that match the grammar in the spec; NO_SUBSTITUTION_TEMPLATE and TEMPLATE_HEAD which starts a template literal, and TEMPLATE_MIDDLE which is between place holder expressions and TEMPLATE_TAIL which ends the template literal.

With this refactoring we don't have to clear lookahead because the scannar takes care of all the relevant whitespace inside the template literal.

Now we can remove clearTokenLookahead.

BUG=None
TEST=None

Review URL: https://codereview.appspot.com/6949052