Change generators to have an @iterator method
For-of loops uses traceur.runtime.getIterator which used to check for the
generatorName symbol. The ES6 spec says to just call @iterator (which is what
SpiderMonkey does) so we add an @iterator method to all generator instances that
just returns the generator instance itself.

BUG=https://code.google.com/p/traceur-compiler/issues/detail?id=190
TEST=No new test

Review URL: https://codereview.appspot.com/7036044