Revert "getopt.js: remove unnecessary 'optdata'."
This reverts commit eee0bfe082aa013478f0f22ded552eb999e27f5b.

Conflicts:
	src/node/getopt.js

The re-indent caused merge conflicts.

The problems with 'optdata' are not really as big as I thought they
were. I don't need to provide accessors or iterators. All 'optdata' can
independently modified, accessed, and iterated over as long as it is
assigned to an object.

Once that is considered, the advantages outweigh the (minimal) extra api
weight.