Add missing `node_modules` dependency to `Makefile`.
`$(TPL_GENSRC_DEPS)` depends on `node_modules`. This was causing `make`
to fail with a fresh repo clone. This isn't the cleanest solution, but
there may not be one, since there is a catch-22 here. We can't know the
full set of prerequisites without at least initializing the
`node_modules` prerequisite.

Discussion link:
https://groups.google.com/d/msg/traceur-compiler-discuss/zWvNP_6gHKg/FlcmRC2MwmsJ