module is not a keyword
This moves the parsing of modules into the fall through case of the big switch of parse statements. If the expression is `module` and it is followed by a module declaration (also definition for now) then it is parsed as a module.

Fixes #353

Review URL: https://github.com/google/traceur-compiler/pull/362

Closes #362, closes #353.