Clean up module code
By adding a top level Module parse tree we can use that to transform
that tree differetntly than a tiop level Script in the inline-module
script.

Remove some dead code now that there is no more identifier base
ModuleDefinitions.

Update the module loader to use `load` and `import` as in the latest
wiki.