Bootstrap System.get/set in runtime.js
We now set up a very rudimentary `System.get` and `System.set` as well as `registerModule` in runtime.js. This does not do any name resolution but that is sufficient to at least work with the output generated by the command line compiler.

This also adds a node test that compiles a js file with an import, prepends runtime.js and the executes that in a new context.

Review URL: https://github.com/google/traceur-compiler/pull/410

Closes #410.