Change code gen for module getters
This changes the code gen for modules to return a simpler object literal. If someone asks for a module instance object, using `module ident from 'name'` or `System.get('name')`, we create the `Module` object at that point.

The benefit for this is that we now have a smaller code gen and it allows follow up changes to dynamically do `export * from 'name'`.

Partial fix for #470

Review URL: https://github.com/google/traceur-compiler/pull/474

Closes #474.