Loader prototype module
Loader.prototype.module no longer takes a name param and we do not register the module in the module registry.

With the old code we got the completion value out of `module` but not any more. The value out was a Module which cannot be toString'ed so I remove the result from the repl.

Also, while I was fixing the repl I fixed a few repl bugs that have been piling up.

Fixes #710
Fixes #713
Fixes #634

Review URL: https://github.com/google/traceur-compiler/pull/716

Closes #716, closes #710, closes #713, closes #634.