only include __moduleName if set
Currently modules with no name are having `__moduleName = (void 0);` written in.

This removes that.

It is useful because a dynamic loader can then set this `__moduleName` variable in an outer scope when evaluating the module to give it a name when loaded.

Review URL: https://api.github.com/repos/google/traceur-compiler/issues/825

Closes #825.