Have SetAccessor use FormalParameterList
Added test in `test/feature/TypeAssertions/ClassParams.js` demonstrates
the #894 issue, which is fixed by this change.

I believe this is the right fix, instead of another solution proposed in #897. SetAccessor has parameterList (FormalParameterList) and therefore
annotations on SetAccessor can be transformed in the same way as
annotations on any regular function.

This change also removes the FormalParameter -> FormalParameterList
transformation in ObjectLiteralTransformer.

Fixes #894
Closes #917, #897