Enable console and log transcoded source for generated-code-dependencies...
When running generated-code-dependencies, the transcoded source and
the traceur-runtime are combined and executed in a new JS sandbox. The
sandbox does not contain the `console` object so console logging fails.

Add flag true to executeFileWithRuntime() in generated-code-dependencies.js
to add the console for logging.

Review URL: https://github.com/google/traceur-compiler/pull/972

Closes #972.