Optimize how we validate cover grammar
When we convert a CoverFormals to a ParenExpression we only need
to check that there is no direct RestParam child.

For the CoverInitializedName we count the seen instances. Using
this we can determine if there have been any since an eralier
point in time and we can skip the expensive tree visitor.