Fixes #1142: select appropriate '@@iterator' in runtime for `for of` statements
[[Work in progress - not for merging]]:

This CL "works" in the browser (Chrome stable, Firefox Nightly) --- http://jsfiddle.net/XDkB6/
(a slightly modified compiled version of one of the failing tests, with options from test-utils.js
and the options from the head of the file), but, as mentioned above, fails two tests in mocha.

Perhaps this can be improved by someone who knows the traceur internals better.