Fix issue with block binding in unbraced for
When transforming the body for if, for, while, do and with, where
either a Block or a single statement is allowed we need to convert
AnonBlocks to normal blocks.

The code for this is generated by the code gen. We use the type
`Block | ParseTree` as the type to trigger this code path.

Fixes #1381