Rework referrer logic to be independent of traceur directory structure.
The overall output of the compiler in multi-file mode is to create a set of
modules in an output file with names beginning with a versioned package name.
The referrer option is encoding both the package name and the path to the
output file from the root of the package directory. EG traceur@0.0.67/bin.
The source is relative to the output directory (eg ../src/Compiler.js).
The referrer causes these relative names to normalize to package names,
eg traceur@0.0.67/src/Compiler. In locate() we map these names back to
the file system using the uncommon part of the referrer and the output
directory (which is the baseURL).

The module names remain unchanged. The referrer now requires a path
and the internal logic that added a traceur-specific path is removed.

Review URL: https://api.github.com/repos/google/traceur-compiler/issues/1419

Closes #1419.