Move ./test/runner.html -> ./testRunner.html.
When run under node, the test system assumes that we start in the root of the Traceur directory.
This assumption is baked into various paths in the tests and expected error messages. To mimic this
assumption in the browser tests we need to move the test file up on directory to the root.
Fixup Loader tests for test/ path.
This is the browser cleanup after removing the 'chdir test'.

Review URL: https://api.github.com/repos/google/traceur-compiler/issues/1555

Closes #1555.