Add System.dirname() and replace __dirname
with System.dirname(__moduleName).

We use `__dirname` in some tests but this value is not part of the standard module environment.
Of course we don't know that `__moduleName` will be standard either, but it seems more likely.

Maybe importing path.dirname would be cleaner, but System.dirname(__moduleName) is
somewhat closer to the convenience of __dirname we replace.

Review URL: https://api.github.com/repos/google/traceur-compiler/issues/1835

Closes #1835.