Fix issue with CommonJS and export star
CommonJS adds the exported bindings as properties to the module.export
object. This is done by transforming the normally exported object into
mutating the module.exports object.

In the case of export star the normally exported object is not an
object literal, but a runtime call. In this case we need to add special
code to extract the relevant data from the call expression.

Fixes #1042