Fixup repl: define System with new BrowserTraceurLoader.
To load the `repl-module.js`, we use dynamic loading, `System.import()`.
The `bin/traceur.js` image contains a `System` only useful for bootloading.
After loading the image on any platform we need to create a loader for
that platform out of the image modules.

To do this for the repl,  we need a reference to `BrowserTraceurLoader`.
This reference has to be via a global. We can use a global `traceur` or
we can use `System.get(<secret path>')`. I picked the first one here.

Extend System.semverMap to compute map from just the version string

Review URL: https://api.github.com/repos/google/traceur-compiler/issues/1855

Closes #1855.