Refactor binary expression parser to reduce call depth
Instead of having to call all of the following:

parseLogicalOR_
parseLogicalAND_
parseBitwiseOR_
parseBitwiseXOR_
parseBitwiseAND_
parseEquality_
parseRelational_
parseShiftExpression_
parseAdditiveExpression_
parseMultiplicativeExpression_
parseExponentiationExpression_
parseUnaryExpression_

We just call parseUnaryExpression_ directly and check if there is
an operator after it. We then use precedence to determine the order
to create the binary expression AST node in.

This gives 10% - 20% performance improvement on the esprima persar
performance test.