Fix issue with missing parens
This is by adding the parens when we see a comma expression in places
where do not expect commas (variable declarations, arrays and calls).

This also adds parens in case the expression of an expression statement
cannot start with the that expression. Currently this is function
expression and object/object pattern. This allows us to generate
cleaner code since we only add the parens for arrow functions as needed.