Add support for parsing JSX
This adds support for Facebook's JSX syntax.

https://facebook.github.io/jsx/

The only thing that is not implemented is JSXSpreadAttribute.

The option to turn this on is called `jsx`. You can provide your own
function to call by setting this option to a string. `--jsx=my.func`.
If not set it defaults to `React.createElement`.