Allow the runtime to be imported as modules
When using `--modules=commonjs --import-runtime` or
`--modules=parse --import-runtime` the generated code will output
require/import statements that import the runtime function as a module.
This means that if you are compiling to commonjs or es6 modules you
can get a minimal runtime and no external runtime file is needed.

Fixes #1628